Update NSComparisonResult to use NS_ENUM
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@40418 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSDictionary.m
@@ -920,7 +920,7 @@ static SEL	appSel;
 
 struct foo { NSDictionary *d; SEL s; IMP i; };
 
-static int
+static NSInteger
 compareIt(id o1, id o2, void* context)
 {
   struct foo	*f = (struct foo*)context;
